fix(mobile): replace activity status indicator icons, ref ENG-37
