+++ packages/ui/native.ts
@@ -1,6 +1,7 @@
 export { Accordion } from './src/components/accordion/accordion.native';
 export { Avatar } from './src/components/avatar/avatar.native';
 export { BulletSeparator } from './src/components/bullet-separator/bullet-separator.native';
+export { ActivityAvatarIcon } from './src/components/avatar/activity-avatar-icon.native';
 export { BtcAvatarIcon } from './src/components/avatar/btc-avatar-icon.native';
 export { RunesAvatarIcon } from './src/components/avatar/runes-avatar-icon.native';
 export { Sip10AvatarIcon } from './src/components/avatar/sip10-avatar-icon.native';
